Added v2 to roles. Builds and tests pass.
