+++ node/libs/roles/src/proto/validator/discovery.proto
@@ -1,0 +1,52 @@
+syntax = "proto3";
+
+package zksync.roles.validator;
+
+import "zksync/std.proto";
+
+// A message broadcasted by a validator
+// over the gossip network announcing
+// its own TCP address. Other validators
+// are expected to establish a consensus
+// network connection to this address.
+message NetAddress {
+  // Address of the validator.
+  optional std.SocketAddr addr = 1; // required
+  // Version of the discovery announcement.
+  // Newer (higher) version overrides the older version.
+  // When a validator gets restarted it broadcasts
+  // its discovery message with version = 0. If it learns
+  // from the network a newer version of its own discovery message
+  // it immediately broadcasts another one to override the received one
+  // (it may happen, since it broadcasted its discovery message also during
+  // the previous execution).
+  //
+  // Currently the IP of the validator is static, but this scheme might also
+  // be used to provide dynamic IP support (if validator discovers that its
+  // own IP has changed - by pinging a trusted STUN server for example - it can
+  // broadcast a new discovery message), or (multi)proxy support (a validator
+  // may maintain a dynamic set of trusted proxy servers which forward traffic
+  // to it - this way validator wouldn't have to have a public IP at all).
+  optional uint64 version = 2; // required
+  // Time at which this message has been signed.
+  // Mostly an informational field:
+  // we cannot use it instead of version field,
+  // because the walltime clock is not guaranteed to be monotone.
+  // We use it as a version "tie breaker" though:
+  // we compare (version,timestamp) lexicographically to
+  // decide which discovery message is newer.
+  // It is useful in the following scenario:
+  // 1. validator broadcasts discovery message with (version1,timestamp1)
+  // 2. validator gets restarted and forgets what it has broadcasted.
+  // 3. validator broadcasts discovery message with (version1,timestamp2)
+  // 4. as long as timestamp1 != timestamp2, one of the messages will be considered
+  //    newer and eventually all nodes will learn this message. It wouldn't be the
+  //    case if both messages had exactly the same version - nodes would keep
+  //    the first message that arrived and decline to broadcast the other.
+  //
+  // Note that although we DO NOT assume timestamps to be monotone,
+  // we assume here that it is unlikely for timestamps to collide.
+  // To make this reasoning more strict, we should rather use a random "tie breaker"
+  // instead (replace timestamp with a random nonce, or use a hash of the entire message).
+  optional std.Timestamp timestamp = 3; // required
+}
