+++ node/libs/roles/src/proto/validator/keys.proto
@@ -16,8 +16,3 @@ message AggregateSignature {
   // TODO: The name is wrong, it should be bls12_381.
   optional bytes bn254 = 1; // required
 }
-
-message WeightedValidator {
-  optional PublicKey key = 1; // required
-  optional uint64 weight = 2; // required
-}
