--- node/libs/roles/src/proto/validator/messages.proto
@@ -1,170 +1,0 @@
-syntax = "proto3";
-
-package zksync.roles.validator;
-
-import "zksync/roles/validator/keys.proto";
-import "zksync/roles/validator/genesis.proto";
-import "zksync/std.proto";
-
-message PayloadHash {
-  optional bytes keccak256 = 1; // required
-}
-
-message BlockHeader {
-  // Sequential number of the block = parent.number + 1.
-  optional uint64 number = 1; // required
-  // Hash of the block payload.
-  optional PayloadHash payload = 2; // required
-}
-
-message FinalBlock {
-  optional bytes payload = 1; // required
-  optional CommitQC justification = 2; // required
-}
-
-message PreGenesisBlock {
-  optional uint64 number = 1; // required
-  optional bytes payload = 2; // required
-  optional bytes justification = 3; // required
-}
-
-message Block {
-  // required
-  oneof t {
-    FinalBlock final = 1;
-    PreGenesisBlock pre_genesis = 2;
-  }
-}
-
-message View {
-  reserved 1,2;
-  reserved "protocol_version","fork";
-
-  optional GenesisHash genesis = 4; // required
-  optional uint64 number = 3; // required; ViewNumber
-}
-
-message ConsensusMsg {
-  reserved 1, 3, 4;
-  reserved "replica_prepare", "leader_prepare", "leader_commit";
-
-  oneof t { // required
-    ReplicaCommit replica_commit = 2;
-    ReplicaTimeout replica_timeout = 5;
-    ReplicaNewView replica_new_view = 6;
-    LeaderProposal leader_proposal = 7;
-  }
-}
-
-message ReplicaCommit {
-  optional View view = 1; // required
-  optional BlockHeader proposal = 2; // required
-}
-
-message ReplicaTimeout {
-  optional View view = 1; // required
-  optional ReplicaCommit high_vote = 2; // optional 
-  optional CommitQC high_qc = 3; // optional
-}
-
-message ReplicaNewView {
-  optional ProposalJustification justification = 1; // required
-}
-
-message LeaderProposal {
-  optional bytes proposal_payload = 1; // optional (depending on justification)
-  optional ProposalJustification justification = 2; // required
-}
-
-message CommitQC {
-  optional ReplicaCommit msg = 1; // required
-  optional std.BitVector signers = 2; // required
-  optional AggregateSignature sig = 3; // required
-}
-
-message TimeoutQC {
-  optional View view = 4; // required
-  repeated ReplicaTimeout msgs = 1; // required
-  repeated std.BitVector signers = 2; // required
-  optional AggregateSignature sig = 3; // required
-}
-
-message ProposalJustification {
-  oneof t { // required
-    CommitQC commit_qc = 1;
-    TimeoutQC timeout_qc = 2;
-  }
-}
-
-message Phase {
-  oneof t { // required
-    std.Void prepare = 1;
-    std.Void commit = 2;
-    std.Void timeout = 3;
-  }
-}
-
-// A message broadcasted by a validator
-// over the gossip network announcing
-// its own TCP address. Other validators
-// are expected to establish a consensus
-// network connection to this address.
-message NetAddress {
-  // Address of the validator.
-  optional std.SocketAddr addr = 1; // required
-  // Version of the discovery announcement.
-  // Newer (higher) version overrides the older version.
-  // When a validator gets restarted it broadcasts
-  // its discovery message with version = 0. If it learns
-  // from the network a newer version of its own discovery message
-  // it immediately broadcasts another one to override the received one
-  // (it may happen, since it broadcasted its discovery message also during
-  // the previous execution).
-  //
-  // Currently the IP of the validator is static, but this scheme might also
-  // be used to provide dynamic IP support (if validator discovers that its
-  // own IP has changed - by pinging a trusted STUN server for example - it can
-  // broadcast a new discovery message), or (multi)proxy support (a validator
-  // may maintain a dynamic set of trusted proxy servers which forward traffic
-  // to it - this way validator wouldn't have to have a public IP at all).
-  optional uint64 version = 2; // required
-  // Time at which this message has been signed.
-  // Mostly an informational field:
-  // we cannot use it instead of version field,
-  // because the walltime clock is not guaranteed to be monotone.
-  // We use it as a version "tie breaker" though:
-  // we compare (version,timestamp) lexicographically to
-  // decide which discovery message is newer.
-  // It is useful in the following scenario:
-  // 1. validator broadcasts discovery message with (version1,timestamp1)
-  // 2. validator gets restarted and forgets what it has broadcasted.
-  // 3. validator broadcasts discovery message with (version1,timestamp2)
-  // 4. as long as timestamp1 != timestamp2, one of the messages will be considered
-  //    newer and eventually all nodes will learn this message. It wouldn't be the
-  //    case if both messages had exactly the same version - nodes would keep
-  //    the first message that arrived and decline to broadcast the other.
-  //
-  // Note that although we DO NOT assume timestamps to be monotone,
-  // we assume here that it is unlikely for timestamps to collide.
-  // To make this reasoning more strict, we should rather use a random "tie breaker"
-  // instead (replace timestamp with a random nonce, or use a hash of the entire message).
-  optional std.Timestamp timestamp = 3; // required
-}
-
-message Msg {
-  oneof t {// required
-    ConsensusMsg consensus = 1;
-    bytes session_id = 2;
-    NetAddress net_address = 3;
-  }
-}
-
-message MsgHash {
-  optional bytes keccak256 = 1; // required
-}
-
-message Signed {
-  optional Msg msg = 1; // required
-  optional PublicKey key = 2; // required
-  optional Signature sig = 3; // required
-}
