--- node/libs/roles/src/validator/conv.rs
@@ -1,351 +1,0 @@
-use std::collections::BTreeMap;
-
-use anyhow::Context as _;
-use zksync_consensus_crypto::ByteFmt;
-use zksync_consensus_utils::enum_util::Variant;
-use zksync_protobuf::{read_optional, read_required, required, ProtoFmt};
-
-use super::{
-    messages::v1::{
-        BlockHeader, CommitQC, FinalBlock, LeaderProposal, LeaderSelectionMode, Phase,
-        ProposalJustification, ReplicaCommit, ReplicaNewView, ReplicaTimeout, Signers, TimeoutQC,
-        View, ViewNumber,
-    },
-    AggregateSignature, Block, BlockNumber, ChainId, Committee, ConsensusMsg, ForkNumber, Genesis,
-    GenesisHash, GenesisRaw, Justification, Msg, MsgHash, NetAddress, Payload, PayloadHash,
-    PreGenesisBlock, ProtocolVersion, PublicKey, Signature, Signed, WeightedValidator,
-};
-use crate::{node::SessionId, proto::validator as proto};
-
-impl ProtoFmt for BlockHeader {
-    type Proto = proto::BlockHeader;
-    fn read(r: &Self::Proto) -> anyhow::Result<Self> {
-        Ok(Self {
-            number: BlockNumber(*required(&r.number).context("number")?),
-            payload: read_required(&r.payload).context("payload")?,
-        })
-    }
-    fn build(&self) -> Self::Proto {
-        Self::Proto {
-            number: Some(self.number.0),
-            payload: Some(self.payload.build()),
-        }
-    }
-}
-
-impl ProtoFmt for FinalBlock {
-    type Proto = proto::FinalBlock;
-    fn read(r: &Self::Proto) -> anyhow::Result<Self> {
-        Ok(Self {
-            payload: Payload(required(&r.payload).context("payload")?.clone()),
-            justification: read_required(&r.justification).context("justification")?,
-        })
-    }
-
-    fn build(&self) -> Self::Proto {
-        Self::Proto {
-            payload: Some(self.payload.0.clone()),
-            justification: Some(self.justification.build()),
-        }
-    }
-}
-
-impl ProtoFmt for View {
-    type Proto = proto::View;
-
-    fn read(r: &Self::Proto) -> anyhow::Result<Self> {
-        Ok(Self {
-            genesis: read_required(&r.genesis).context("genesis")?,
-            number: ViewNumber(*required(&r.number).context("number")?),
-        })
-    }
-
-    fn build(&self) -> Self::Proto {
-        Self::Proto {
-            genesis: Some(self.genesis.build()),
-            number: Some(self.number.0),
-        }
-    }
-}
-
-impl ProtoFmt for ReplicaCommit {
-    type Proto = proto::ReplicaCommit;
-
-    fn read(r: &Self::Proto) -> anyhow::Result<Self> {
-        Ok(Self {
-            view: read_required(&r.view).context("view")?,
-            proposal: read_required(&r.proposal).context("proposal")?,
-        })
-    }
-
-    fn build(&self) -> Self::Proto {
-        Self::Proto {
-            view: Some(self.view.build()),
-            proposal: Some(self.proposal.build()),
-        }
-    }
-}
-
-impl ProtoFmt for ReplicaTimeout {
-    type Proto = proto::ReplicaTimeout;
-
-    fn read(r: &Self::Proto) -> anyhow::Result<Self> {
-        Ok(Self {
-            view: read_required(&r.view).context("view")?,
-            high_vote: read_optional(&r.high_vote).context("high_vote")?,
-            high_qc: read_optional(&r.high_qc).context("high_qc")?,
-        })
-    }
-
-    fn build(&self) -> Self::Proto {
-        Self::Proto {
-            view: Some(self.view.build()),
-            high_vote: self.high_vote.as_ref().map(ProtoFmt::build),
-            high_qc: self.high_qc.as_ref().map(ProtoFmt::build),
-        }
-    }
-}
-
-impl ProtoFmt for ReplicaNewView {
-    type Proto = proto::ReplicaNewView;
-
-    fn read(r: &Self::Proto) -> anyhow::Result<Self> {
-        Ok(Self {
-            justification: read_required(&r.justification).context("justification")?,
-        })
-    }
-
-    fn build(&self) -> Self::Proto {
-        Self::Proto {
-            justification: Some(self.justification.build()),
-        }
-    }
-}
-
-impl ProtoFmt for LeaderProposal {
-    type Proto = proto::LeaderProposal;
-
-    fn read(r: &Self::Proto) -> anyhow::Result<Self> {
-        Ok(Self {
-            proposal_payload: r.proposal_payload.as_ref().map(|p| Payload(p.clone())),
-            justification: read_required(&r.justification).context("justification")?,
-        })
-    }
-
-    fn build(&self) -> Self::Proto {
-        Self::Proto {
-            proposal_payload: self.proposal_payload.as_ref().map(|p| p.0.clone()),
-            justification: Some(self.justification.build()),
-        }
-    }
-}
-
-impl ProtoFmt for CommitQC {
-    type Proto = proto::CommitQc;
-
-    fn read(r: &Self::Proto) -> anyhow::Result<Self> {
-        Ok(Self {
-            message: read_required(&r.msg).context("msg")?,
-            signers: read_required(&r.signers).context("signers")?,
-            signature: read_required(&r.sig).context("sig")?,
-        })
-    }
-
-    fn build(&self) -> Self::Proto {
-        Self::Proto {
-            msg: Some(self.message.build()),
-            signers: Some(self.signers.build()),
-            sig: Some(self.signature.build()),
-        }
-    }
-}
-
-impl ProtoFmt for TimeoutQC {
-    type Proto = proto::TimeoutQc;
-
-    fn read(r: &Self::Proto) -> anyhow::Result<Self> {
-        let mut map = BTreeMap::new();
-
-        for (msg, signers) in r.msgs.iter().zip(r.signers.iter()) {
-            map.insert(
-                ReplicaTimeout::read(msg).context("msg")?,
-                Signers::read(signers).context("signers")?,
-            );
-        }
-
-        Ok(Self {
-            view: read_required(&r.view).context("view")?,
-            map,
-            signature: read_required(&r.sig).context("sig")?,
-        })
-    }
-
-    fn build(&self) -> Self::Proto {
-        let (msgs, signers) = self
-            .map
-            .iter()
-            .map(|(msg, signers)| (msg.build(), signers.build()))
-            .unzip();
-
-        Self::Proto {
-            view: Some(self.view.build()),
-            msgs,
-            signers,
-            sig: Some(self.signature.build()),
-        }
-    }
-}
-
-impl ProtoFmt for ProposalJustification {
-    type Proto = proto::ProposalJustification;
-
-    fn read(r: &Self::Proto) -> anyhow::Result<Self> {
-        use proto::proposal_justification::T;
-        Ok(match r.t.as_ref().context("missing")? {
-            T::CommitQc(r) => Self::Commit(ProtoFmt::read(r).context("Commit")?),
-            T::TimeoutQc(r) => Self::Timeout(ProtoFmt::read(r).context("Timeout")?),
-        })
-    }
-
-    fn build(&self) -> Self::Proto {
-        use proto::proposal_justification::T;
-
-        let t = match self {
-            Self::Commit(x) => T::CommitQc(x.build()),
-            Self::Timeout(x) => T::TimeoutQc(x.build()),
-        };
-
-        Self::Proto { t: Some(t) }
-    }
-}
-
-impl ProtoFmt for Signers {
-    type Proto = zksync_protobuf::proto::std::BitVector;
-
-    fn read(r: &Self::Proto) -> anyhow::Result<Self> {
-        Ok(Self(ProtoFmt::read(r)?))
-    }
-
-    fn build(&self) -> Self::Proto {
-        self.0.build()
-    }
-}
-
-impl ProtoFmt for Phase {
-    type Proto = proto::Phase;
-
-    fn read(r: &Self::Proto) -> anyhow::Result<Self> {
-        use proto::phase::T;
-        Ok(match required(&r.t)? {
-            T::Prepare(_) => Self::Prepare,
-            T::Commit(_) => Self::Commit,
-            T::Timeout(_) => Self::Timeout,
-        })
-    }
-
-    fn build(&self) -> Self::Proto {
-        use proto::phase::T;
-        let t = match self {
-            Self::Prepare => T::Prepare(zksync_protobuf::proto::std::Void {}),
-            Self::Commit => T::Commit(zksync_protobuf::proto::std::Void {}),
-            Self::Timeout => T::Timeout(zksync_protobuf::proto::std::Void {}),
-        };
-        Self::Proto { t: Some(t) }
-    }
-}
-
-impl ProtoFmt for PublicKey {
-    type Proto = proto::PublicKey;
-
-    fn read(r: &Self::Proto) -> anyhow::Result<Self> {
-        Ok(Self(ByteFmt::decode(required(&r.bn254)?)?))
-    }
-
-    fn build(&self) -> Self::Proto {
-        Self::Proto {
-            bn254: Some(self.0.encode()),
-        }
-    }
-}
-
-impl ProtoFmt for Signature {
-    type Proto = proto::Signature;
-
-    fn read(r: &Self::Proto) -> anyhow::Result<Self> {
-        Ok(Self(ByteFmt::decode(required(&r.bn254)?)?))
-    }
-
-    fn build(&self) -> Self::Proto {
-        Self::Proto {
-            bn254: Some(self.0.encode()),
-        }
-    }
-}
-
-impl ProtoFmt for LeaderSelectionMode {
-    type Proto = proto::LeaderSelectionMode;
-
-    fn read(r: &Self::Proto) -> anyhow::Result<Self> {
-        match required(&r.mode)? {
-            proto::leader_selection_mode::Mode::RoundRobin(_) => {
-                Ok(LeaderSelectionMode::RoundRobin)
-            }
-            proto::leader_selection_mode::Mode::Sticky(inner) => {
-                let key = required(&inner.key).context("key")?;
-                Ok(LeaderSelectionMode::Sticky(PublicKey::read(key)?))
-            }
-            proto::leader_selection_mode::Mode::Weighted(_) => Ok(LeaderSelectionMode::Weighted),
-            proto::leader_selection_mode::Mode::Rota(inner) => {
-                let _ = required(&inner.keys.first()).context("keys")?;
-                let pks = inner
-                    .keys
-                    .iter()
-                    .map(PublicKey::read)
-                    .collect::<Result<Vec<_>, _>>()?;
-                Ok(LeaderSelectionMode::Rota(pks))
-            }
-        }
-    }
-    fn build(&self) -> Self::Proto {
-        match self {
-            LeaderSelectionMode::RoundRobin => proto::LeaderSelectionMode {
-                mode: Some(proto::leader_selection_mode::Mode::RoundRobin(
-                    proto::leader_selection_mode::RoundRobin {},
-                )),
-            },
-            LeaderSelectionMode::Sticky(pk) => proto::LeaderSelectionMode {
-                mode: Some(proto::leader_selection_mode::Mode::Sticky(
-                    proto::leader_selection_mode::Sticky {
-                        key: Some(pk.build()),
-                    },
-                )),
-            },
-            LeaderSelectionMode::Weighted => proto::LeaderSelectionMode {
-                mode: Some(proto::leader_selection_mode::Mode::Weighted(
-                    proto::leader_selection_mode::Weighted {},
-                )),
-            },
-            LeaderSelectionMode::Rota(pks) => proto::LeaderSelectionMode {
-                mode: Some(proto::leader_selection_mode::Mode::Rota(
-                    proto::leader_selection_mode::Rota {
-                        keys: pks.iter().map(|pk| pk.build()).collect(),
-                    },
-                )),
-            },
-        }
-    }
-}
-
-impl ProtoFmt for AggregateSignature {
-    type Proto = proto::AggregateSignature;
-
-    fn read(r: &Self::Proto) -> anyhow::Result<Self> {
-        Ok(Self(ByteFmt::decode(required(&r.bn254)?)?))
-    }
-
-    fn build(&self) -> Self::Proto {
-        Self::Proto {
-            bn254: Some(self.0.encode()),
-        }
-    }
-}
