+++ node/libs/roles/src/validator/keys/public_key.rs
@@ -2,6 +2,9 @@ use std::fmt;
 
 use zksync_consensus_crypto::{bls12_381, ByteFmt, Text, TextFmt};
 
+use crate::proto::validator as proto;
+use zksync_protobuf::{required, ProtoFmt};
+
 /// A public key for a validator.
 #[derive(Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct PublicKey(pub(crate) bls12_381::PublicKey);
@@ -26,6 +29,20 @@ impl TextFmt for PublicKey {
         text.strip("validator:public:bls12_381:")?
             .decode_hex()
             .map(Self)
+    }
+}
+
+impl ProtoFmt for PublicKey {
+    type Proto = proto::PublicKey;
+
+    fn read(r: &Self::Proto) -> anyhow::Result<Self> {
+        Ok(Self(ByteFmt::decode(required(&r.bn254)?)?))
+    }
+
+    fn build(&self) -> Self::Proto {
+        Self::Proto {
+            bn254: Some(self.0.encode()),
+        }
     }
 }
