+++ node/libs/roles/src/validator/messages/v1/consensus.rs
@@ -7,6 +7,10 @@ use zksync_consensus_crypto::keccak256::Keccak256;
 
 use crate::validator::{Committee, Genesis, GenesisHash, PublicKey};
 
+use crate::proto::validator as proto;
+use anyhow::Context as _;
+use zksync_protobuf::{read_required, required, ProtoFmt};
+
 /// View specification.
 /// WARNING: any change to this struct may invalidate preexisting signatures. See `TimeoutQC` docs.
 #[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
@@ -38,6 +42,24 @@ impl View {
             genesis: self.genesis,
             number,
         })
+    }
+}
+
+impl ProtoFmt for View {
+    type Proto = proto::View;
+
+    fn read(r: &Self::Proto) -> anyhow::Result<Self> {
+        Ok(Self {
+            genesis: read_required(&r.genesis).context("genesis")?,
+            number: ViewNumber(*required(&r.number).context("number")?),
+        })
+    }
+
+    fn build(&self) -> Self::Proto {
+        Self::Proto {
+            genesis: Some(self.genesis.build()),
+            number: Some(self.number.0),
+        }
     }
 }
 
@@ -121,6 +143,60 @@ impl LeaderSelectionMode {
     }
 }
 
+impl ProtoFmt for LeaderSelectionMode {
+    type Proto = proto::LeaderSelectionMode;
+
+    fn read(r: &Self::Proto) -> anyhow::Result<Self> {
+        match required(&r.mode)? {
+            proto::leader_selection_mode::Mode::RoundRobin(_) => {
+                Ok(LeaderSelectionMode::RoundRobin)
+            }
+            proto::leader_selection_mode::Mode::Sticky(inner) => {
+                let key = required(&inner.key).context("key")?;
+                Ok(LeaderSelectionMode::Sticky(PublicKey::read(key)?))
+            }
+            proto::leader_selection_mode::Mode::Weighted(_) => Ok(LeaderSelectionMode::Weighted),
+            proto::leader_selection_mode::Mode::Rota(inner) => {
+                let _ = required(&inner.keys.first()).context("keys")?;
+                let pks = inner
+                    .keys
+                    .iter()
+                    .map(PublicKey::read)
+                    .collect::<Result<Vec<_>, _>>()?;
+                Ok(LeaderSelectionMode::Rota(pks))
+            }
+        }
+    }
+    fn build(&self) -> Self::Proto {
+        match self {
+            LeaderSelectionMode::RoundRobin => proto::LeaderSelectionMode {
+                mode: Some(proto::leader_selection_mode::Mode::RoundRobin(
+                    proto::leader_selection_mode::RoundRobin {},
+                )),
+            },
+            LeaderSelectionMode::Sticky(pk) => proto::LeaderSelectionMode {
+                mode: Some(proto::leader_selection_mode::Mode::Sticky(
+                    proto::leader_selection_mode::Sticky {
+                        key: Some(pk.build()),
+                    },
+                )),
+            },
+            LeaderSelectionMode::Weighted => proto::LeaderSelectionMode {
+                mode: Some(proto::leader_selection_mode::Mode::Weighted(
+                    proto::leader_selection_mode::Weighted {},
+                )),
+            },
+            LeaderSelectionMode::Rota(pks) => proto::LeaderSelectionMode {
+                mode: Some(proto::leader_selection_mode::Mode::Rota(
+                    proto::leader_selection_mode::Rota {
+                        keys: pks.iter().map(|pk| pk.build()).collect(),
+                    },
+                )),
+            },
+        }
+    }
+}
+
 /// Struct that represents a bit map of validators. We use it to compactly store
 /// which validators signed a given message.
 /// WARNING: any change to this struct may invalidate preexisting signatures. See `TimeoutQC` docs.
@@ -163,6 +239,18 @@ impl Signers {
     }
 }
 
+impl ProtoFmt for Signers {
+    type Proto = zksync_protobuf::proto::std::BitVector;
+
+    fn read(r: &Self::Proto) -> anyhow::Result<Self> {
+        Ok(Self(ProtoFmt::read(r)?))
+    }
+
+    fn build(&self) -> Self::Proto {
+        self.0.build()
+    }
+}
+
 impl std::ops::BitOrAssign<&Self> for Signers {
     fn bitor_assign(&mut self, other: &Self) {
         self.0.or(&other.0);
@@ -191,4 +279,27 @@ pub enum Phase {
     Prepare,
     Commit,
     Timeout,
+}
+
+impl ProtoFmt for Phase {
+    type Proto = proto::Phase;
+
+    fn read(r: &Self::Proto) -> anyhow::Result<Self> {
+        use proto::phase::T;
+        Ok(match required(&r.t)? {
+            T::Prepare(_) => Self::Prepare,
+            T::Commit(_) => Self::Commit,
+            T::Timeout(_) => Self::Timeout,
+        })
+    }
+
+    fn build(&self) -> Self::Proto {
+        use proto::phase::T;
+        let t = match self {
+            Self::Prepare => T::Prepare(zksync_protobuf::proto::std::Void {}),
+            Self::Commit => T::Commit(zksync_protobuf::proto::std::Void {}),
+            Self::Timeout => T::Timeout(zksync_protobuf::proto::std::Void {}),
+        };
+        Self::Proto { t: Some(t) }
+    }
 }
