+++ node/libs/roles/src/validator/messages/v1/replica_new_view.rs
@@ -1,6 +1,10 @@
 use super::{ProposalJustification, ProposalJustificationVerifyError, View};
 use crate::validator::Genesis;
 
+use crate::proto::validator as proto;
+use anyhow::Context as _;
+use zksync_protobuf::{read_required, ProtoFmt};
+
 /// A new view message from a replica.
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct ReplicaNewView {
@@ -22,6 +26,22 @@ impl ReplicaNewView {
             .map_err(ReplicaNewViewVerifyError::Justification)?;
 
         Ok(())
+    }
+}
+
+impl ProtoFmt for ReplicaNewView {
+    type Proto = proto::ReplicaNewView;
+
+    fn read(r: &Self::Proto) -> anyhow::Result<Self> {
+        Ok(Self {
+            justification: read_required(&r.justification).context("justification")?,
+        })
+    }
+
+    fn build(&self) -> Self::Proto {
+        Self::Proto {
+            justification: Some(self.justification.build()),
+        }
     }
 }
