+++ node/libs/roles/src/validator/messages/v2/tests/replica_commit.rs
@@ -3,11 +3,11 @@ use rand::Rng;
 use zksync_concurrency::ctx;
 
 use super::*;
-use crate::validator::{messages::tests::genesis_v1, testonly::Setup, ChainId, Signed};
+use crate::validator::{messages::tests::genesis_v2, testonly::Setup, ChainId, Signed};
 
 #[test]
 fn test_replica_commit_verify() {
-    let mut genesis = genesis_v1();
+    let mut genesis = genesis_v2();
     let commit = replica_commit();
     assert!(commit.verify(&genesis).is_ok());
 
@@ -26,7 +26,7 @@ fn test_commit_qc_add() {
     let rng = &mut ctx.rng();
     let setup = Setup::new(rng, 2);
     let view = rng.gen();
-    let mut qc = CommitQC::new(setup.make_replica_commit_v1(rng, view), &setup.genesis);
+    let mut qc = CommitQC::new(setup.make_replica_commit_v2(rng, view), &setup.genesis);
     let msg = qc.message.clone();
 
     // Add the first signature
@@ -103,7 +103,7 @@ fn test_commit_qc_verify() {
     let rng = &mut ctx.rng();
     let setup = Setup::new(rng, 6);
     let view = rng.gen();
-    let qc = setup.make_commit_qc_v1(rng, view);
+    let qc = setup.make_commit_qc_v2(rng, view);
 
     // Verify the QC
     assert!(qc.verify(&setup.genesis).is_ok());
